# packages/api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Koupreng API
  version: 0.1.0
  description: Verified route skeleton from Spring controller annotations. Schemas are incomplete where DTO details were not fully mapped.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/invitations/{invitationId}/guests:
    get:
      summary: List guests
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/invitationId"
      responses:
        "200":
          description: Insufficient data to verify response schema.
    post:
      summary: Create guest
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/invitationId"
      responses:
        "201":
          description: Insufficient data to verify response schema.
...
  /api/v1/invitations/{invitationId}/gifts:
    get:
      summary: List wedding gifts
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/invitationId"
      responses:
        "200":
          description: Insufficient data to verify response schema.
    post:
      summary: Create wedding gift
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/invitationId"
      responses:
        "200":
          description: Insufficient data to verify response schema.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    invitationId:
      name: invitationId
      in: path
      required: true
      schema:
        type: integer
        format: int64